backend/requirements.txt
# MuseumVoice V3 - Système Structuré d'Œuvres d'Art

# === OBLIGATOIRES ===
# API Web
flask>=3.1.0
flask-cors>=6.0.0
gunicorn>=21.2.0

# Base de données PostgreSQL
psycopg2-binary>=2.9.9

# Traitement PDF intelligent
PyPDF2>=3.0.0

# Base scientifique
numpy>=1.23.0
scikit-learn>=1.0.0

# Utilitaires
requests>=2.28.0

# === RECOMMANDÉES (haute performance) ===
# Embeddings et recherche vectorielle pour RAG
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2

# Traitement avancé de texte
torch>=2.0.0

# Sécurité
bcrypt>=4.0.0

# === OPTIONNELLES (fonctionnalités avancées) ===
# LLM APIs
groq>=0.4.0
openai>=1.0.0

# Framework RAG avancé
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Interface utilisateur
# streamlit>=1.28.0
# gradio>=3.50.0

# Base de données avancée
# chromadb>=0.4.0
# pgvector>=0.2.0

# Monitoring avancé
# prometheus-client>=0.17.0
# psutil>=5.9.0